Split Is* predicate helpers out of PInvokeGenerator.cs
Move the contiguous cluster of type/statement inspection predicates
(IsEnumOperator through IsUnsigned, including IsExcluded, IsFixedSize,
IsType, IsUnchecked, IsUnsafe, and the IsTransparentStruct* helpers) into
a dedicated PInvokeGenerator.Predicates.cs partial. Trims the main partial
by ~1590 lines. Pure code move; no behavior change.
